requirements.txt
pandas
psycopg
apache-airflow
datetime
requests
pandasql
urllib3
traceback2
matplotlib